add searchengine class
